Add api handlers for GET /contacts and GET /contacts/{id}
We are using Gorilla Mux package here because the ServeMux in the
standard library doesn't provide an easy way to bind different handlers
based on the request method. If we were to use the servemux of the
standard library, we have to write a switch case statement for every
pattern match and then specify a handler. Using gorilla/mux makes this
easy and also provides an easy way to get parameters from the URL, such as
{id} in /contacts/{id}.